Restructure to Align `Speak` and `Listen` Modules
