+++ deepgram/clients/speak/v1/rest/__init__.py
@@ -1,0 +1,7 @@
+# Copyright 2024 Deepgram SDK contributors. All Rights Reserved.
+# Use of this source code is governed by a MIT license that can be found in the LICENSE file.
+# SPDX-License-Identifier: MIT
+
+from .client import SpeakRESTClient
+from .async_client import AsyncSpeakRESTClient
+from .response import SpeakRESTResponse
