+++ deepgram/clients/speak/v1/websocket/__init__.py
@@ -2,14 +2,9 @@
 # Use of this source code is governed by a MIT license that can be found in the LICENSE file.
 # SPDX-License-Identifier: MIT
 
-from .options import SpeakOptions, FileSource, SpeakWebSocketSource, SpeakSource
-from ....options import DeepgramClientOptions, ClientOptionsFromEnv
-
-from .rest import SpeakRESTClient, AsyncSpeakRESTClient
-from .rest import SpeakRESTResponse
-
-from .websocket import SpeakWebSocketClient, AsyncSpeakWebSocketClient
-from .websocket import (
+from .client import SpeakWebSocketClient
+from .async_client import AsyncSpeakWebSocketClient
+from .response import (
     SpeakWebSocketResponse,
     OpenResponse,
     MetadataResponse,
