+++ deepgram/clients/speak/v1/websocket/async_client.py
@@ -1,21 +1,20 @@
 # Copyright 2024 Deepgram SDK contributors. All Rights Reserved.
 # Use of this source code is governed by a MIT license that can be found in the LICENSE file.
 # SPDX-License-Identifier: MIT
-
+import asyncio
 import json
-import time
 import logging
 from typing import Dict, Union, Optional, cast, Any
 import threading
 
-from websockets.sync.client import connect, ClientConnection
 import websockets
+from websockets.client import WebSocketClientProtocol
 
-from ....utils import verboselogs
-from ....options import DeepgramClientOptions
-from ..enums import SpeakStreamEvents
-from ...live.helpers import convert_to_websocket_url, append_query_params
-from ..errors import DeepgramError
+from .....utils import verboselogs
+from .....options import DeepgramClientOptions
+from ...enums import SpeakWebSocketEvents
+from ....live.helpers import convert_to_websocket_url, append_query_params
+from ...errors import DeepgramError
 
 from .response import (
     OpenResponse,
@@ -26,10 +25,10 @@ from .response import (
     ErrorResponse,
     UnhandledResponse,
 )
-from .options import SpeakOptions
+from ..options import SpeakOptions
 
 
-class SpeakStreamClient:  # pylint: disable=too-many-instance-attributes
+class AsyncSpeakWebSocketClient:  # pylint: disable=too-many-instance-attributes
     """
     Client for interacting with Deepgram's text-to-speech services over WebSockets.
 
@@ -44,13 +43,10 @@ class SpeakStreamClient:  # pylint: disable=too-many-instance-attributes
     _endpoint: str
     _websocket_url: str
 
-    _socket: ClientConnection
-    _exit_event: threading.Event
-    _lock_send: threading.Lock
-    _event_handlers: Dict[SpeakStreamEvents, list]
+    _socket: WebSocketClientProtocol
+    _event_handlers: Dict[SpeakWebSocketEvents, list]
 
-
-    _listen_thread: Union[threading.Thread, None]
+    _listen_thread: Union[asyncio.Task, None]
 
     _kwargs: Optional[Dict] = None
     _addons: Optional[Dict] = None
@@ -67,20 +63,19 @@ class SpeakStreamClient:  # pylint: disable=too-many-instance-attributes
 
         self._config = config
         self._endpoint = "v1/speak"
-        self._lock_send = threading.Lock()
 
         self._listen_thread = None
 
         # exit
-        self._exit_event = threading.Event()
+        self._exit_event = asyncio.Event()
 
         self._event_handlers = {
-            event: [] for event in SpeakStreamEvents.__members__.values()
+            event: [] for event in SpeakWebSocketEvents.__members__.values()
         }
         self._websocket_url = convert_to_websocket_url(self._config.url, self._endpoint)
 
-    # pylint: disable=too-many-statements,too-many-branches
-    def start(
+    # pylint: disable=too-many-branches,too-many-statements
+    async def start(
         self,
         options: Optional[Union[SpeakOptions, Dict]] = None,
         addons: Optional[Dict] = None,
@@ -91,7 +86,7 @@ class SpeakStreamClient:  # pylint: disable=too-many-instance-attributes
         """
         Starts the WebSocket connection for text-to-speech synthesis.
         """
-        self._logger.debug("SpeakStreamClient.start ENTER")
+        self._logger.debug("AsyncSpeakStreamClient.start ENTER")
         self._logger.info("options: %s", options)
         self._logger.info("addons: %s", addons)
         self._logger.info("headers: %s", headers)
@@ -139,8 +134,11 @@ class SpeakStreamClient:  # pylint: disable=too-many-instance-attributes
         self._logger.debug("combined_headers: %s", combined_headers)
 
         url_with_params = append_query_params(self._websocket_url, combined_options)
+
         try:
-            self._socket = connect(url_with_params, additional_headers=combined_headers)
+            self._socket = await websockets.connect(
+                url_with_params, extra_headers=combined_headers
+            )
             self._exit_event.clear()
 
             # debug the threads
@@ -148,9 +146,8 @@ class SpeakStreamClient:  # pylint: disable=too-many-instance-attributes
                 self._logger.debug("after running thread: %s", thread.name)
             self._logger.debug("number of active threads: %s", threading.active_count())
 
-            # listening thread
-            self._listen_thread = threading.Thread(target=self._listening)
-            self._listen_thread.start()
+            # listen thread
+            self._listen_thread = asyncio.create_task(self._listening())
 
             # debug the threads
             for thread in threading.enumerate():
@@ -158,144 +155,173 @@ class SpeakStreamClient:  # pylint: disable=too-many-instance-attributes
             self._logger.debug("number of active threads: %s", threading.active_count())
 
             # push open event
-            self._emit(
-                SpeakStreamEvents(SpeakStreamEvents.Open),
-                OpenResponse(type=SpeakStreamEvents.Open),
+            await self._emit(
+                SpeakWebSocketEvents(SpeakWebSocketEvents.Open),
+                OpenResponse(type=SpeakWebSocketEvents.Open),
             )
 
             self._logger.notice("start succeeded")
-            self._logger.debug("SpeakStreamClient.start LEAVE")
+            self._logger.debug("AsyncSpeakStreamClient.start LEAVE")
             return True
         except websockets.ConnectionClosed as e:
-            self._logger.error("ConnectionClosed in SpeakStreamClient.start: %s", e)
-            self._logger.debug("SpeakStreamClient.start LEAVE")
+            self._logger.error(
+                "ConnectionClosed in AsyncSpeakStreamClient.start: %s", e
+            )
+            self._logger.debug("AsyncSpeakStreamClient.start LEAVE")
             if self._config.options.get("termination_exception_connect") == "true":
-                raise e
+                raise
             return False
         except websockets.exceptions.WebSocketException as e:
-            self._logger.error("WebSocketException in SpeakStreamClient.start: %s", e)
-            self._logger.debug("SpeakStreamClient.start LEAVE")
+            self._logger.error(
+                "WebSocketException in AsyncSpeakStreamClient.start: %s", e
+            )
+            self._logger.debug("AsyncSpeakStreamClient.start LEAVE")
             if self._config.options.get("termination_exception_connect") == "true":
-                raise e
+                raise
             return False
         except Exception as e:  # pylint: disable=broad-except
-            self._logger.error("WebSocketException in SpeakStreamClient.start: %s", e)
-            self._logger.debug("SpeakStreamClient.start LEAVE")
+            self._logger.error(
+                "WebSocketException in AsyncSpeakStreamClient.start: %s", e
+            )
+            self._logger.debug("AsyncSpeakStreamClient.start LEAVE")
             if self._config.options.get("termination_exception_connect") == "true":
-                raise e
+                raise
             return False
 
-    # pylint: enable=too-many-statements,too-many-branches
+    # pylint: enable=too-many-branches,too-many-statements
 
-    def on(
-        self, event: SpeakStreamEvents, handler
-    ) -> None:  # registers event handlers for specific events
+    def on(self, event: SpeakWebSocketEvents, handler) -> None:
         """
         Registers event handlers for specific events.
         """
         self._logger.info("event subscribed: %s", event)
-        if event in SpeakStreamEvents.__members__.values() and callable(handler):
-            self._event_handlers[event].append(handler)
+        if event in SpeakWebSocketEvents.__members__.values() and callable(handler):
+            if handler not in self._event_handlers[event]:
+                self._event_handlers[event].append(handler)
 
-    def _emit(self, event: SpeakStreamEvents, *args, **kwargs) -> None:
+    # triggers the registered event handlers for a specific event
+    async def _emit(self, event: SpeakWebSocketEvents, *args, **kwargs) -> None:
         """
         Emits events to the registered event handlers.
         """
+        self._logger.debug("AsyncSpeakStreamClient._emit ENTER")
         self._logger.debug("callback handlers for: %s", event)
+
+        # debug the threads
+        for thread in threading.enumerate():
+            self._logger.debug("after running thread: %s", thread.name)
+        self._logger.debug("number of active threads: %s", threading.active_count())
+
+        tasks = []
         for handler in self._event_handlers[event]:
-            handler(self, *args, **kwargs)
+            task = asyncio.create_task(handler(self, *args, **kwargs))
+            tasks.append(task)
+
+        if tasks:
+            self._logger.debug("waiting for tasks to finish...")
+            await asyncio.gather(*tasks, return_exceptions=True)
+            tasks.clear()
+
+        # debug the threads
+        for thread in threading.enumerate():
+            self._logger.debug("after running thread: %s", thread.name)
+        self._logger.debug("number of active threads: %s", threading.active_count())
+
+        self._logger.debug("AsyncSpeakStreamClient._emit LEAVE")
 
     # pylint: disable=too-many-return-statements,too-many-statements,too-many-locals,too-many-branches
-    def _listening(
-        self,
-    ) -> None:
+    async def _listening(self) -> None:
         """
         Listens for messages from the WebSocket connection.
         """
-        self._logger.debug("SpeakStreamClient._listening ENTER")
+        self._logger.debug("AsyncSpeakStreamClient._listening ENTER")
 
         while True:
             try:
                 if self._exit_event.is_set():
                     self._logger.notice("_listening exiting gracefully")
-                    self._logger.debug("SpeakStreamClient._listening LEAVE")
+                    self._logger.debug("AsyncSpeakStreamClient._listening LEAVE")
                     return
 
                 if self._socket is None:
                     self._logger.warning("socket is empty")
-                    self._logger.debug("SpeakStreamClient._listening LEAVE")
+                    self._logger.debug("AsyncSpeakStreamClient._listening LEAVE")
                     return
 
-                message = self._socket.recv()
+                message = await self._socket.recv()
 
                 if message is None:
-                    self._logger.info("message is empty")
+                    self._logger.spam("message is None")
                     continue
-                
+
                 if isinstance(message, bytes):
                     self._logger.debug("Binary data received")
-                    self._emit(
-                        SpeakStreamEvents(SpeakStreamEvents.AudioData),
+                    await self._emit(
+                        SpeakWebSocketEvents(SpeakWebSocketEvents.AudioData),
                         data=message,
                         **dict(cast(Dict[Any, Any], self._kwargs)),
-                        )
+                    )
                 else:
                     data = json.loads(message)
                     response_type = data.get("type")
-                    self._logger.debug("response_type: %s, data: %s", response_type, data)
+                    self._logger.debug(
+                        "response_type: %s, data: %s", response_type, data
+                    )
 
                     match response_type:
-                        case SpeakStreamEvents.Open:
+                        case SpeakWebSocketEvents.Open:
                             open_result: OpenResponse = OpenResponse.from_json(message)
                             self._logger.verbose("OpenResponse: %s", open_result)
-                            self._emit(
-                                SpeakStreamEvents(SpeakStreamEvents.Open),
+                            await self._emit(
+                                SpeakWebSocketEvents(SpeakWebSocketEvents.Open),
                                 open=open_result,
                                 **dict(cast(Dict[Any, Any], self._kwargs)),
                             )
-                        case SpeakStreamEvents.Metadata:
+                        case SpeakWebSocketEvents.Metadata:
                             meta_result: MetadataResponse = MetadataResponse.from_json(
                                 message
                             )
                             self._logger.verbose("MetadataResponse: %s", meta_result)
-                            self._emit(
-                                SpeakStreamEvents(SpeakStreamEvents.Metadata),
+                            await self._emit(
+                                SpeakWebSocketEvents(SpeakWebSocketEvents.Metadata),
                                 metadata=meta_result,
                                 **dict(cast(Dict[Any, Any], self._kwargs)),
                             )
-                        case SpeakStreamEvents.Flush:
-                            fl_result: FlushedResponse = (
-                                FlushedResponse.from_json(message)
+                        case SpeakWebSocketEvents.Flush:
+                            fl_result: FlushedResponse = FlushedResponse.from_json(
+                                message
                             )
                             self._logger.verbose("FlushedResponse: %s", fl_result)
-                            self._emit(
-                                SpeakStreamEvents(
-                                    SpeakStreamEvents.Flush
-                                ),
+                            await self._emit(
+                                SpeakWebSocketEvents(SpeakWebSocketEvents.Flush),
                                 flushed=fl_result,
                                 **dict(cast(Dict[Any, Any], self._kwargs)),
                             )
-                        case SpeakStreamEvents.Close:
-                            close_result: CloseResponse = CloseResponse.from_json(message)
+                        case SpeakWebSocketEvents.Close:
+                            close_result: CloseResponse = CloseResponse.from_json(
+                                message
+                            )
                             self._logger.verbose("CloseResponse: %s", close_result)
-                            self._emit(
-                                SpeakStreamEvents(SpeakStreamEvents.Close),
+                            await self._emit(
+                                SpeakWebSocketEvents(SpeakWebSocketEvents.Close),
                                 close=close_result,
                                 **dict(cast(Dict[Any, Any], self._kwargs)),
                             )
-                        case SpeakStreamEvents.Warning:
-                            war_warning: WarningResponse = WarningResponse.from_json(message)
+                        case SpeakWebSocketEvents.Warning:
+                            war_warning: WarningResponse = WarningResponse.from_json(
+                                message
+                            )
                             self._logger.verbose("WarningResponse: %s", war_warning)
-                            self._emit(
-                                SpeakStreamEvents(SpeakStreamEvents.Warning),
+                            await self._emit(
+                                SpeakWebSocketEvents(SpeakWebSocketEvents.Warning),
                                 warning=war_warning,
                                 **dict(cast(Dict[Any, Any], self._kwargs)),
                             )
-                        case SpeakStreamEvents.Error:
+                        case SpeakWebSocketEvents.Error:
                             err_error: ErrorResponse = ErrorResponse.from_json(message)
                             self._logger.verbose("ErrorResponse: %s", err_error)
-                            self._emit(
-                                SpeakStreamEvents(SpeakStreamEvents.Error),
+                            await self._emit(
+                                SpeakWebSocketEvents(SpeakWebSocketEvents.Error),
                                 error=err_error,
                                 **dict(cast(Dict[Any, Any], self._kwargs)),
                             )
@@ -306,46 +332,48 @@ class SpeakStreamClient:  # pylint: disable=too-many-instance-attributes
                                 data,
                             )
                             unhandled_error: UnhandledResponse = UnhandledResponse(
-                                type=SpeakStreamEvents(
-                                    SpeakStreamEvents.Unhandled
+                                type=SpeakWebSocketEvents(
+                                    SpeakWebSocketEvents.Unhandled
                                 ),
                                 raw=message,
                             )
-                            self._emit(
-                                SpeakStreamEvents(SpeakStreamEvents.Unhandled),
+                            await self._emit(
+                                SpeakWebSocketEvents(SpeakWebSocketEvents.Unhandled),
                                 unhandled=unhandled_error,
                                 **dict(cast(Dict[Any, Any], self._kwargs)),
                             )
 
             except websockets.exceptions.ConnectionClosedOK as e:
                 self._logger.notice(f"_listening({e.code}) exiting gracefully")
-                self._logger.debug("SpeakStreamClient._listening LEAVE")
+                self._logger.debug("AsyncSpeakStreamClient._listening LEAVE")
                 return
 
             except websockets.exceptions.ConnectionClosed as e:
                 if e.code == 1000:
                     self._logger.notice(f"_listening({e.code}) exiting gracefully")
-                    self._logger.debug("SpeakStreamClient._listening LEAVE")
+                    self._logger.debug("AsyncSpeakStreamClient._listening LEAVE")
                     return
 
                 self._logger.error(
-                    "ConnectionClosed in SpeakStreamClient._listening with code %s: %s",
+                    "ConnectionClosed in AsyncSpeakStreamClient._listening with code %s: %s",
                     e.code,
                     e.reason,
                 )
                 cc_error: ErrorResponse = ErrorResponse(
-                    "ConnectionClosed in SpeakStreamClient._listening",
+                    "ConnectionClosed in AsyncSpeakStreamClient._listening",
                     f"{e}",
                     "ConnectionClosed",
                 )
-                self._emit(
-                    SpeakStreamEvents(SpeakStreamEvents.Error), cc_error
+                await self._emit(
+                    SpeakWebSocketEvents(SpeakWebSocketEvents.Error),
+                    error=cc_error,
+                    **dict(cast(Dict[Any, Any], self._kwargs)),
                 )
 
                 # signal exit and close
-                self._signal_exit()
+                await self._signal_exit()
 
-                self._logger.debug("SpeakStreamClient._listening LEAVE")
+                self._logger.debug("AsyncSpeakStreamClient._listening LEAVE")
 
                 if self._config.options.get("termination_exception") == "true":
                     raise
@@ -353,123 +381,126 @@ class SpeakStreamClient:  # pylint: disable=too-many-instance-attributes
 
             except websockets.exceptions.WebSocketException as e:
                 self._logger.error(
-                    "WebSocketException in SpeakStreamClient._listening with: %s", e
+                    "WebSocketException in AsyncSpeakStreamClient._listening: %s", e
                 )
                 ws_error: ErrorResponse = ErrorResponse(
-                    "WebSocketException in SpeakStreamClient._listening",
+                    "WebSocketException in AsyncSpeakStreamClient._listening",
                     f"{e}",
                     "WebSocketException",
                 )
-                self._emit(
-                    SpeakStreamEvents(SpeakStreamEvents.Error), ws_error
+                await self._emit(
+                    SpeakWebSocketEvents(SpeakWebSocketEvents.Error),
+                    error=ws_error,
+                    **dict(cast(Dict[Any, Any], self._kwargs)),
                 )
 
                 # signal exit and close
-                self._signal_exit()
+                await self._signal_exit()
 
-                self._logger.debug("SpeakStreamClient._listening LEAVE")
+                self._logger.debug("AsyncSpeakStreamClient._listening LEAVE")
 
                 if self._config.options.get("termination_exception") == "true":
                     raise
                 return
 
             except Exception as e:  # pylint: disable=broad-except
-                self._logger.error("Exception in SpeakStreamClient._listening: %s", e)
+                self._logger.error(
+                    "Exception in AsyncSpeakStreamClient._listening: %s", e
+                )
                 e_error: ErrorResponse = ErrorResponse(
-                    "Exception in SpeakStreamClient._listening",
+                    "Exception in AsyncSpeakStreamClient._listening",
                     f"{e}",
                     "Exception",
                 )
-                self._logger.error("Exception in SpeakStreamClient._listening: %s", str(e))
-                self._emit(
-                    SpeakStreamEvents(SpeakStreamEvents.Error), e_error
+                await self._emit(
+                    SpeakWebSocketEvents(SpeakWebSocketEvents.Error),
+                    error=e_error,
+                    **dict(cast(Dict[Any, Any], self._kwargs)),
                 )
 
                 # signal exit and close
-                self._signal_exit()
+                await self._signal_exit()
 
-                self._logger.debug("SpeakStreamClient._listening LEAVE")
+                self._logger.debug("AsyncSpeakStreamClient._listening LEAVE")
 
                 if self._config.options.get("termination_exception") == "true":
                     raise
                 return
 
-
     # pylint: disable=too-many-return-statements
-    def send(self, text_input: str) -> bool:
+
+    async def send(self, text_input: str) -> bool:
         """
         Sends data over the WebSocket connection.
         """
-        self._logger.spam("SpeakStreamClient.send ENTER")
+        self._logger.spam("AsyncSpeakStreamClient.send ENTER")
 
         if self._exit_event.is_set():
             self._logger.notice("send exiting gracefully")
-            self._logger.debug("SpeakStreamClient.send LEAVE")
+            self._logger.debug("AsyncSpeakStreamClient.send LEAVE")
             return False
 
         if self._socket is not None:
-            with self._lock_send:
-                try:
-                    self._socket.send(json.dumps({"type": "Speak", "text": text_input}))
-                except websockets.exceptions.ConnectionClosedOK as e:
-                    self._logger.notice(f"send() exiting gracefully: {e.code}")
-                    self._logger.debug("SpeakStreamClient._keep_alive LEAVE")
+            try:
+                await self._socket.send(
+                    json.dumps({"type": "Speak", "text": text_input})
+                )
+            except websockets.exceptions.ConnectionClosedOK as e:
+                self._logger.notice(f"send() exiting gracefully: {e.code}")
+                self._logger.debug("AsyncSpeakStreamClient.send LEAVE")
+                if self._config.options.get("termination_exception_send") == "true":
+                    raise
+                return True
+            except websockets.exceptions.ConnectionClosed as e:
+                if e.code in [1000, 1001]:
+                    self._logger.notice(f"send({e.code}) exiting gracefully")
+                    self._logger.debug("AsyncSpeakStreamClient.send LEAVE")
                     if self._config.options.get("termination_exception_send") == "true":
                         raise
                     return True
-                except websockets.exceptions.ConnectionClosed as e:
-                    if e.code == 1000:
-                        self._logger.notice(f"send({e.code}) exiting gracefully")
-                        self._logger.debug("SpeakStreamClient.send LEAVE")
-                        if (
-                            self._config.options.get("termination_exception_send")
-                            == "true"
-                        ):
-                            raise
-                        return True
-                    self._logger.error("send() failed - ConnectionClosed: %s", str(e))
-                    self._logger.spam("SpeakStreamClient.send LEAVE")
-                    if self._config.options.get("termination_exception_send") == "true":
-                        raise
-                    return False
-                except websockets.exceptions.WebSocketException as e:
-                    self._logger.error("send() failed - WebSocketException: %s", str(e))
-                    self._logger.spam("SpeakStreamClient.send LEAVE")
-                    if self._config.options.get("termination_exception_send") == "true":
-                        raise
-                    return False
-                except Exception as e:  # pylint: disable=broad-except
-                    self._logger.error("send() failed - Exception: %s", str(e))
-                    self._logger.spam("SpeakStreamClient.send LEAVE")
-                    if self._config.options.get("termination_exception_send") == "true":
-                        raise
-                    return False
+
+                self._logger.error("send() failed - ConnectionClosed: %s", str(e))
+                self._logger.spam("AsyncSpeakStreamClient.send LEAVE")
+                if self._config.options.get("termination_exception_send") == "true":
+                    raise
+                return False
+            except websockets.exceptions.WebSocketException as e:
+                self._logger.error("send() failed - WebSocketException: %s", str(e))
+                self._logger.spam("AsyncSpeakStreamClient.send LEAVE")
+                if self._config.options.get("termination_exception_send") == "true":
+                    raise
+                return False
+            except Exception as e:  # pylint: disable=broad-except
+                self._logger.error("send() failed - Exception: %s", str(e))
+                self._logger.spam("AsyncSpeakStreamClient.send LEAVE")
+                if self._config.options.get("termination_exception_send") == "true":
+                    raise
+                return False
 
             self._logger.spam("send() succeeded")
-            self._logger.spam("SpeakStreamClient.send LEAVE")
+            self._logger.spam("AsyncSpeakStreamClient.send LEAVE")
             return True
 
         self._logger.spam("send() failed. socket is None")
-        self._logger.spam("SpeakStreamClient.send LEAVE")
+        self._logger.spam("AsyncSpeakStreamClient.send LEAVE")
         return False
 
     # pylint: enable=too-many-return-statements
 
-
-    def flush(self) -> bool:
+    async def flush(self) -> bool:
         """
         Flushes the current buffer and returns generated audio
         """
-        self._logger.spam("SpeakStreamClient.flush ENTER")
+        self._logger.spam("AsyncSpeakStreamClient.flush ENTER")
 
         if self._exit_event.is_set():
             self._logger.notice("flush exiting gracefully")
-            self._logger.debug("SpeakStreamClient.flush LEAVE")
+            self._logger.debug("AsyncSpeakStreamClient.flush LEAVE")
             return False
 
         if self._socket is None:
             self._logger.notice("socket is not intialized")
-            self._logger.debug("SpeakStreamClient.flush LEAVE")
+            self._logger.debug("AsyncSpeakStreamClient.flush LEAVE")
             return False
 
         self._logger.notice("Sending Flush...")
@@ -477,58 +508,76 @@ class SpeakStreamClient:  # pylint: disable=too-many-instance-attributes
 
         if not ret:
             self._logger.error("flush failed")
-            self._logger.spam("SpeakStreamClient.flush LEAVE")
+            self._logger.spam("AsyncSpeakStreamClient.flush LEAVE")
             return False
 
         self._logger.notice("flush succeeded")
-        self._logger.spam("SpeakStreamClient.flush LEAVE")
+        self._logger.spam("AsyncSpeakStreamClient.flush LEAVE")
 
         return True
 
-    # closes the WebSocket connection gracefully
-    def finish(self) -> bool:
+    async def finish(self) -> bool:
         """
         Closes the WebSocket connection gracefully.
         """
-        self._logger.spam("SpeakStreamClient.finish ENTER")
-
-        # debug the threads
-        for thread in threading.enumerate():
-            self._logger.debug("before running thread: %s", thread.name)
-        self._logger.debug("number of active threads: %s", threading.active_count())
+        self._logger.debug("AsyncSpeakStreamClient.finish ENTER")
 
         # signal exit
-        self._signal_exit()
+        await self._signal_exit()
 
         # stop the threads
+        self._logger.verbose("cancelling tasks...")
+        try:
+            # Before cancelling, check if the tasks were created
+            # debug the threads
+            for thread in threading.enumerate():
+                self._logger.debug("before running thread: %s", thread.name)
+            self._logger.debug("number of active threads: %s", threading.active_count())
 
-        if self._listen_thread is not None:
-            self._listen_thread.join()
-            self._listen_thread = None
-        self._logger.notice("listening thread joined")
+            tasks = []
 
-        # debug the threads
-        for thread in threading.enumerate():
-            self._logger.debug("before running thread: %s", thread.name)
-        self._logger.debug("number of active threads: %s", threading.active_count())
+            if self._listen_thread is not None:
+                self._listen_thread.cancel()
+                tasks.append(self._listen_thread)
+                self._logger.notice("processing _listen_thread cancel...")
 
-        self._logger.notice("finish succeeded")
-        self._logger.spam("SpeakStreamClient.finish LEAVE")
-        return True
+            # Use asyncio.gather to wait for tasks to be cancelled
+            await asyncio.wait_for(
+                asyncio.gather(*tasks), timeout=10
+            )  # Prevent indefinite waiting
+            self._logger.notice("threads joined")
 
-    # signals the WebSocket connection to exit
-    def _signal_exit(self) -> None:
-        # closes the WebSocket connection gracefully
-        self._logger.notice("closing socket...")
+            # debug the threads
+            for thread in threading.enumerate():
+                self._logger.debug("after running thread: %s", thread.name)
+            self._logger.debug("number of active threads: %s", threading.active_count())
+
+            self._logger.notice("finish succeeded")
+            self._logger.spam("AsyncSpeakStreamClient.finish LEAVE")
+            return True
+
+        except asyncio.CancelledError as e:
+            self._logger.error("tasks cancelled error: %s", e)
+            self._logger.debug("AsyncSpeakStreamClient.finish LEAVE")
+            return False
+
+        except asyncio.TimeoutError as e:
+            self._logger.error("tasks cancellation timed out: %s", e)
+            self._logger.debug("AsyncSpeakStreamClient.finish LEAVE")
+            return False
+
+    async def _signal_exit(self) -> None:
+        # send close event
+        self._logger.verbose("closing socket...")
         if self._socket is not None:
-            self._logger.notice("sending Close...")
+            self._logger.verbose("send CloseStream...")
             try:
                 # if the socket connection is closed, the following line might throw an error
-                self._socket.send(json.dumps({"type": "Close"}))
+                await self.send(json.dumps({"type": "CloseStream"}))
             except websockets.exceptions.ConnectionClosedOK as e:
                 self._logger.notice("_signal_exit  - ConnectionClosedOK: %s", e.code)
             except websockets.exceptions.ConnectionClosed as e:
-                self._logger.error("_signal_exit  - ConnectionClosed: %s", e.code)
+                self._logger.notice("_signal_exit  - ConnectionClosed: %s", e.code)
             except websockets.exceptions.WebSocketException as e:
                 self._logger.error("_signal_exit - WebSocketException: %s", str(e))
             except Exception as e:  # pylint: disable=broad-except
@@ -536,15 +585,16 @@ class SpeakStreamClient:  # pylint: disable=too-many-instance-attributes
 
             # push close event
             try:
-                self._emit(
-                    SpeakStreamEvents(SpeakStreamEvents.Close),
-                    CloseResponse(type=SpeakStreamEvents.Close),
+                await self._emit(
+                    SpeakWebSocketEvents(SpeakWebSocketEvents.Close),
+                    close=CloseResponse(type=SpeakWebSocketEvents.Close),
+                    **dict(cast(Dict[Any, Any], self._kwargs)),
                 )
             except Exception as e:  # pylint: disable=broad-except
-                self._logger.error("_signal_exit - Exception: %s", e)
+                self._logger.error("_emit - Exception: %s", e)
 
             # wait for task to send
-            time.sleep(0.5)
+            await asyncio.sleep(0.5)
 
         # signal exit
         self._exit_event.set()
@@ -554,7 +604,7 @@ class SpeakStreamClient:  # pylint: disable=too-many-instance-attributes
         if self._socket is not None:
             self._logger.verbose("socket.wait_closed...")
             try:
-                self._socket.close()
+                await self._socket.close()
             except websockets.exceptions.WebSocketException as e:
                 self._logger.error("socket.wait_closed failed: %s", e)
